Scrape Updates
Changed the name of the output file to something that makes a bit more
sense.
Updated the scrape, so that it now deletes null named entries in the
output, and if a module is taught with another (and as such will not
include the lesson data), it will find the correct module with the
correct (hopefully) lesson data and copy it over.
Cleaned up a few other bits so that it makes a bit more sense now.