Fixed Groups
Scraping the groups of a Prac was broken, is now mostly fixed. It's
pretty buggy at the minute, but hopefully I'll be able to improve it to
make it more reliable. There are too many modules for me to be definite
that it's 100% working, so for now I'm going to assume it is.

A backup timetable for Semester 1 2014 was also added, just for fun